backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
email-validator